+++ cordformation/src/main/kotlin/net/corda/plugins/Cordform.kt
@@ -62,7 +62,6 @@ open class Cordform : Baseform() {
         nodes.forEach(Node::installConfig)
         installCordaJar()
         generateExcludedWhitelist()
-        generateIncludeWhitelist()
         generateKeystoreAndSignCordappJar()
         installRunScript()
         nodes.forEach(Node::installDrivers)